Add InstallerHelperTests.cs Add ProductTests.cs Add VideoServiceTests.cs Add EmployeeControllerTests.cs

+++ Awesome-Automated-Test/Mocking/Employee/EmployeeController.cs
@@ -4,24 +4,46 @@ namespace Awesome_Automated_Test.Mocking.Employee
 {
     public class EmployeeController
     {
-        private EmployeeContext _db;
+        private readonly IEmployeeStorage _employeeStorage;
 
-        public EmployeeController()
+        public EmployeeController(IEmployeeStorage employeeStorage)
         {
-            _db = new EmployeeContext();
+            _employeeStorage = employeeStorage;
         }
 
         public ActionResult DeleteEmployee(int id)
         {
-            var employee = _db.Employees.Find(id);
-            _db.Employees.Remove(employee);
-            _db.SaveChanges();
+            _employeeStorage.DeleteEmployee(id);
             return RedirectToAction("Employees");
         }
 
         private ActionResult RedirectToAction(string employees)
         {
             return new RedirectResult();
+        }
+    }
+
+    public interface IEmployeeStorage
+    {
+        void DeleteEmployee(int id);
+    }
+
+    public class EmployeeStorage : IEmployeeStorage
+    {
+        private readonly EmployeeContext _db;
+
+        public EmployeeStorage()
+        {
+            _db = new EmployeeContext();
+        }
+
+        public void DeleteEmployee(int id)
+        {
+            var employee = _db.Employees.Find(id);
+            if (employee == null) return;
+
+            _db.Employees.Remove(employee);
+            _db.SaveChanges();
         }
     }
 }
